test(ci): shrink release preflight runner fixture
